requirements.txt
kafka-python==2.3.0
websocket-client==1.8.0
